services/auth/requirements.txt
grpcio==1.76.0
grpcio-tools==1.76.0
grpcio-reflection==1.76.0

sqlalchemy[asyncio]==2.0.46
asyncpg==0.31.0

pydantic==2.12.5
pydantic-settings==2.12.0

bcrypt==4.2.1
PyJWT==2.10.1

python-dotenv==1.2.1

pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-mock==3.15.1
